# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenZeppelin Relayer API",
...
    "termsOfService": "https://www.openzeppelin.com/tos",
    "contact": {
      "name": "OpenZeppelin",
      "url": "https://www.openzeppelin.com"
    },
    "license": {
      "name": "AGPL-3.0 license",
      "url": "https://github.com/OpenZeppelin/openzeppelin-relayer/blob/main/LICENSE"
    },
    "version": "1.1.0"
  },
  "paths": {
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Notification routes implementation",
        "description": "Note: OpenAPI documentation for these endpoints can be found in the `openapi.rs` file\n\nLists all notifications with pagination support.",
        "operationId": "listNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (starts at 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_NotificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Bad Request",
                  "success": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Unauthorized",
                  "success": false
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Internal Server Error",
                  "success": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-required-permissions": {
          "permissions": [
            "notifications:read"
          ],
          "scope": "global"
        }
      },
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Creates a new notification.",
        "operationId": "createNotification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NotificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Bad Request",
                  "success": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Unauthorized",
                  "success": false
                }
              }
            }
          },
          "409": {
            "description": "Notification with this ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Notification with this ID already exists",
                  "success": false
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Internal Server Error",
                  "success": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-required-permissions": {
          "permissions": [
            "notifications:create"
          ],
          "scope": "global"
        }
      }
    },
...
    "/api/v1/relayers": {
      "get": {
        "tags": [
          "Relayers"
        ],
        "summary": "Relayer routes implementation",
        "description": "Note: OpenAPI documentation for these endpoints can be found in the `openapi.rs` file\n\nLists all relayers with pagination support.",
        "operationId": "listRelayers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (starts at 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relayer list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_RelayerResponse"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Bad Request",
                  "success": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Unauthorized",
                  "success": false
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Too Many Requests",
                  "success": false
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Internal Server Error",
                  "success": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-required-permissions": {
          "permissions": [
            "relayers:read"
          ],
          "scope": "global"
        }
      },
      "post": {
        "tags": [
          "Relayers"
        ],
        "summary": "Creates a new relayer.",
        "operationId": "createRelayer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelayerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Relayer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RelayerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Bad Request",
                  "success": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Unauthorized",
                  "success": false
                }
              }
            }
          },
          "409": {
            "description": "Relayer with this ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Relayer with this ID already exists",
                  "success": false
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Internal Server Error",
                  "success": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-required-permissions": {
          "permissions": [
            "relayers:create"
          ],
          "scope": "global"
        }
      }
    },
...
    "/api/v1/signers": {
      "get": {
        "tags": [
          "Signers"
        ],
        "summary": "Signer routes implementation",
        "description": "Note: OpenAPI documentation for these endpoints can be found in the `openapi.rs` file\n\nLists all signers with pagination support.",
        "operationId": "listSigners",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (starts at 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signer list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_SignerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Bad Request",
                  "success": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Unauthorized",
                  "success": false
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Internal Server Error",
                  "success": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-required-permissions": {
          "permissions": [
            "signers:read"
          ],
          "scope": "global"
        }
      },
      "post": {
        "tags": [
          "Signers"
        ],
        "summary": "Creates a new signer.",
        "operationId": "createSigner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Signer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SignerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Bad Request",
                  "success": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Unauthorized",
                  "success": false
                }
              }
            }
          },
          "409": {
            "description": "Signer with this ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Signer with this ID already exists",
                  "success": false
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                },
                "example": {
                  "data": null,
                  "message": "Internal Server Error",
                  "success": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-required-permissions": {
          "permissions": [
            "signers:create"
          ],
          "scope": "global"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_NotificationResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Response structure for notification API endpoints",
            "required": [
              "id",
              "type",
              "url",
              "has_signing_key"
            ],
            "properties": {
              "has_signing_key": {
                "type": "boolean",
                "description": "Signing key is hidden in responses for security"
              },
              "id": {
                "type": "string"
              },
              "type": {
                "$ref": "#/components/schemas/NotificationType"
              },
              "url": {
                "type": "string"
              }
            }
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_RelayerResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "Relayer response model for API endpoints",
            "required": [
              "id",
              "name",
              "network",
              "network_type",
              "paused",
              "signer_id"
            ],
            "properties": {
              "address": {
                "type": "string"
              },
              "custom_rpc_urls": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RpcConfig"
                }
              },
              "disabled_reason": {
                "$ref": "#/components/schemas/DisabledReason"
              },
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "network": {
                "type": "string"
              },
              "network_type": {
                "$ref": "#/components/schemas/RelayerNetworkType"
              },
              "notification_id": {
                "type": "string"
              },
              "paused": {
                "type": "boolean"
              },
              "policies": {
                "$ref": "#/components/schemas/RelayerNetworkPolicyResponse",
                "description": "Policies without redundant network_type tag - network type is available at top level\nOnly included if user explicitly provided policies (not shown for empty/default policies)"
              },
              "signer_id": {
                "type": "string"
              },
              "system_disabled": {
                "type": "boolean"
              }
            }
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_SignerResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "type",
              "config"
            ],
            "properties": {
              "config": {
                "$ref": "#/components/schemas/SignerConfigResponse",
                "description": "Non-secret configuration details"
              },
              "id": {
                "type": "string",
                "description": "The unique identifier of the signer"
              },
              "type": {
                "$ref": "#/components/schemas/SignerType",
                "description": "The type of signer (local, aws_kms, google_cloud_kms, vault, etc.)"
              }
            }
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_String": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_Vec_NotificationResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Response structure for notification API endpoints",
              "required": [
                "id",
                "type",
                "url",
                "has_signing_key"
              ],
              "properties": {
                "has_signing_key": {
                  "type": "boolean",
                  "description": "Signing key is hidden in responses for security"
                },
                "id": {
                  "type": "string"
                },
                "type": {
                  "$ref": "#/components/schemas/NotificationType"
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_Vec_RelayerResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Relayer response model for API endpoints",
              "required": [
                "id",
                "name",
                "network",
                "network_type",
                "paused",
                "signer_id"
              ],
              "properties": {
                "address": {
                  "type": "string"
                },
                "custom_rpc_urls": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RpcConfig"
                  }
                },
                "disabled_reason": {
                  "$ref": "#/components/schemas/DisabledReason"
                },
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "network": {
                  "type": "string"
                },
                "network_type": {
                  "$ref": "#/components/schemas/RelayerNetworkType"
                },
                "notification_id": {
                  "type": "string"
                },
                "paused": {
                  "type": "boolean"
                },
                "policies": {
                  "$ref": "#/components/schemas/RelayerNetworkPolicyResponse",
                  "description": "Policies without redundant network_type tag - network type is available at top level\nOnly included if user explicitly provided policies (not shown for empty/default policies)"
                },
                "signer_id": {
                  "type": "string"
                },
                "system_disabled": {
                  "type": "boolean"
                }
              }
            }
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_Vec_SignerResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "type",
                "config"
              ],
              "properties": {
                "config": {
                  "$ref": "#/components/schemas/SignerConfigResponse",
                  "description": "Non-secret configuration details"
                },
                "id": {
                  "type": "string",
                  "description": "The unique identifier of the signer"
                },
                "type": {
                  "$ref": "#/components/schemas/SignerType",
                  "description": "The type of signer (local, aws_kms, google_cloud_kms, vault, etc.)"
                }
              }
            }
          },
          "error": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/PluginMetadata"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "CreateRelayerRequest": {
        "type": "object",
        "description": "Request model for creating a new relayer",
        "required": [
          "name",
          "network",
          "paused",
          "network_type",
          "signer_id"
        ],
        "properties": {
          "custom_rpc_urls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpcConfig"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "network": {
            "type": "string"
          },
          "network_type": {
            "$ref": "#/components/schemas/RelayerNetworkType"
          },
          "notification_id": {
            "type": "string"
          },
          "paused": {
            "type": "boolean"
          },
          "policies": {
            "$ref": "#/components/schemas/CreateRelayerPolicyRequest",
            "description": "Policies - will be deserialized based on the network_type field"
          },
          "signer_id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "NotificationCreateRequest": {
        "type": "object",
        "description": "Request structure for creating a new notification",
        "required": [
          "url"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "signing_key": {
            "type": "string",
            "description": "Optional signing key for securing webhook notifications"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "url": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "NotificationType": {
        "type": "string",
        "description": "Notification type enum used by both config file and API",
        "enum": [
          "webhook"
        ]
      },
...
      "PaginationMeta": {
        "type": "object",
        "required": [
          "current_page",
          "per_page",
          "total_items"
        ],
        "properties": {
          "current_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PluginMetadata": {
        "type": "object",
        "properties": {
          "logs": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/LogEntry"
            }
          },
          "traces": {
            "type": [
              "array",
              "null"
            ],
            "items": {}
          }
        }
      },
...
      "RelayerNetworkPolicyResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/EvmPolicyResponse"
          },
          {
            "$ref": "#/components/schemas/StellarPolicyResponse"
          },
          {
            "$ref": "#/components/schemas/SolanaPolicyResponse"
          }
        ],
        "description": "Policy types for responses - these don't include network_type tags\nsince the network_type is already available at the top level of RelayerResponse"
      },
      "RelayerNetworkType": {
        "type": "string",
        "description": "Network type enum for relayers",
        "enum": [
          "evm",
          "solana",
          "stellar"
        ]
      },
...
      "RpcConfig": {
        "type": "object",
        "description": "Configuration for an RPC endpoint.",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "The RPC endpoint URL."
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "description": "The weight of this endpoint in the weighted round-robin selection.\nDefaults to DEFAULT_RPC_WEIGHT (100). Should be between 0 and 100.",
            "minimum": 0
          }
        }
      },
...
      "SignerCreateRequest": {
        "type": "object",
        "description": "Request model for creating a new signer",
        "required": [
          "type",
          "config"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/SignerConfigRequest",
            "description": "The signer configuration"
          },
          "id": {
            "type": "string",
            "description": "Optional ID - if not provided, a UUID will be generated"
          },
          "type": {
            "$ref": "#/components/schemas/SignerTypeRequest",
            "description": "The type of signer"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}